Logger optimization part 3
