+++ common/chrono/clock/clock.hpp
@@ -5,7 +5,6 @@
 #include <string_view>
 
 namespace demiplane::chrono {
-
     // ─────────────── Base clock (parse & now) ───────────────
     class Clock {
     public:
@@ -17,7 +16,8 @@ namespace demiplane::chrono {
 
         template <class Dur = std::chrono::milliseconds>
         [[nodiscard]] static std::chrono::time_point<std::chrono::system_clock, Dur> parse(
-            const std::string_view txt, std::string_view fmt) {
+            const std::string_view txt,
+            std::string_view fmt) {
             using namespace std::chrono;
             std::istringstream is(std::string{txt});
             sys_time<Dur> tp;
@@ -47,14 +47,53 @@ namespace demiplane::chrono {
         [[nodiscard]] static std::string current_time(const std::string_view fmt) {
             return format_time(now(), fmt);
         }
+
         static std::string format_time_iso_ms(const sys_tp tp) {
-            const auto ms    = std::chrono::duration_cast<std::chrono::milliseconds>(tp.time_since_epoch()) % 1000;
-            std::string base = format_time(tp, clock_formats::iso8601);
-            base += std::format(".{:03}", ms.count());
+            const auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(
+                                tp.time_since_epoch()) % 1000;
+            const auto tm = to_tm(tp);
+            char timestamp[32];
+            std::int32_t len;
             if constexpr (CT == ClockType::UTC) {
-                base += 'Z';
+                len = snprintf(timestamp, sizeof(timestamp),
+                               "%04d-%02d-%02dT%02d:%02d:%02d.%03ldZ",
+                               tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday,
+                               tm.tm_hour, tm.tm_min, tm.tm_sec,
+                               ms.count());
             }
-            return base;
+            else {
+                len = snprintf(timestamp, sizeof(timestamp),
+                               "%04d-%02d-%02dT%02d:%02d:%02d.%03ld ",
+                               tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday,
+                               tm.tm_hour, tm.tm_min, tm.tm_sec,
+                               ms.count());
+            }
+
+            return {timestamp, static_cast<std::size_t>(len)};
+        }
+
+        static void format_time_iso_ms(const sys_tp tp, std::string& out) {
+            const auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(
+                                tp.time_since_epoch()) % 1000;
+            const auto tm = to_tm(tp);
+            char timestamp[32];
+            std::int32_t len;
+            if constexpr (CT == ClockType::UTC) {
+                len = snprintf(timestamp, sizeof(timestamp),
+                               "%04d-%02d-%02dT%02d:%02d:%02d.%03ldZ",
+                               tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday,
+                               tm.tm_hour, tm.tm_min, tm.tm_sec,
+                               ms.count());
+            }
+            else {
+                len = snprintf(timestamp, sizeof(timestamp),
+                               "%04d-%02d-%02dT%02d:%02d:%02d.%03ld ",
+                               tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday,
+                               tm.tm_hour, tm.tm_min, tm.tm_sec,
+                               ms.count());
+            }
+
+            out.append(timestamp, static_cast<std::size_t>(len));
         }
         [[nodiscard]] static std::string format_time(const sys_tp tp, const std::string_view format) {
             std::ostringstream ss;
@@ -62,12 +101,14 @@ namespace demiplane::chrono {
             ss << std::put_time(&buf, format.data());
             return ss.str();
         }
+
         [[nodiscard]] static std::tm to_tm(const sys_tp tp) {
             const std::time_t tt = std::chrono::system_clock::to_time_t(tp);
             std::tm out{};
             if constexpr (CT == ClockType::Local) {
                 localtime_r(&tt, &out);
-            } else {
+            }
+            else {
                 gmtime_r(&tt, &out);
             }
             return out;
@@ -76,5 +117,4 @@ namespace demiplane::chrono {
 
     using LocalClock = SpecClock<ClockType::Local>;
     using UTCClock   = SpecClock<ClockType::UTC>;
-
 } // namespace demiplane::chrono
