+++ common/scroll/entry/include/factory/entry_factory.hpp
@@ -1,12 +1,11 @@
 #pragma once
 
-#include "entry/custom_entry.hpp"
-#include "entry/detailed_entry.hpp"
-#include "entry/fast_detailed_entry.hpp"
-#include "entry/light_entry.hpp"
-#include "entry/service_entry.hpp"
-namespace demiplane::scroll {
+#include "custom_entry.hpp"
+#include "detailed_entry.hpp"
+#include "light_entry.hpp"
+#include "service_entry.hpp"
 
+namespace demiplane::scroll {
     template <class EntryT, class... Extra>
     EntryT make_entry(LogLevel lvl,
                       const std::string_view msg,
