+++ common/scroll/entry/include/service_entry.hpp
@@ -2,9 +2,8 @@
 
 #include <sstream>
 #include <string>
-
+//todo: make private
 #include <clock.hpp>
-#include <gears_concepts.hpp>
 #include "../entry_interface.hpp"
 
 namespace demiplane::scroll {
